+++ app/wppwebjs/session/health.ts
@@ -1,5 +1,5 @@
 import type { WAState } from "whatsapp-web.js";
-import { HEALTH_CHECK_INTERVAL_MS } from "./config";
+import { HEALTH_CHECK_INTERVAL_MS } from ".";
 import type { HealthResult } from "../interfaces";
 import { safeDestroyClient, scheduleReconnect } from "./handlers";
 import { sessions } from ".";
