--- app/wppwebjs/whatsapp.ts
@@ -1,7 +1,0 @@
-export * from "./session";
-export * from "./session/client";
-export * from "./session/status";
-export * from "./session/health";
-export * from "./number-utils";
-export * from "./messaging";
-export * from "./session/management";
